feat: added Example Long-Running intent implementation - reverted initial lib example to it's first responsibility - extracted AppShortcuts in example to be able to register new intents - updated Readme
